Rename packages and clean
